Replace LeadingStrings benchmarks with rebar Industry benchmarks
Port curated/01-literal and curated/02-literal-alternate benchmarks from
https://github.com/BurntSushi/rebar for Russian and Chinese text.
Haystacks are OpenSubtitles data (https://opus.nlpl.eu/OpenSubtitles-v2018.php).

This replaces the ad-hoc NonAscii benchmark with well-established cross-engine
regex benchmarks that test literal and alternation search on non-ASCII text.
